+++ test/unit/lib/fs-utils.spec.ts
